requirements.txt
Flask
azure-identity
azure-keyvault-secrets
openai
azure-cosmos
python-dotenv
